Add Record#rdata_txt to each record class
Adds a rdata_txt method to every record that returns the record's rdata
formatted as it would be in a zone file. Google's Cloud DNS API is largely
based on the formatting of string data (which to be fair is how a zone file
works so I don't blame them). Exposing a method that formats rdata as an
RFC compliant string makes it much easier to work with the API.

With the exposure of rdata_txt, a refactor of Record#to_s present
itself, allowing us to cut down on a substantial amount of code by
hoisting to_s logic into the Record class. Given the re-use of Record
instances, I created an adhoc fixtures hash. If it's deemed useful, a future
refactor can extract the records into a YAML file.